test_case/requirements.txt
pytest==7.4.3
allure-pytest==2.13.2
requests==2.31.0
pytest-html==4.1.1